Add answers to question previews
This is smart: if it's multiple choice, it shows all the answers, with
any marked correct with a check in the box. If the question is a free
response question, then it will show the answer marked correct. If no
answer is marked correct, it will just show the first one.